В общем что-то не так.
Побитово сверил с примером в задании FileMap. Совпадает. Если не
сработает в колыбельке -- не знаю что делать. Далее. С MultiFileMap.
Колыбелька хочет ошибку, если нет такой папки, которая нужна. Буквально
на этом семинаре вы просили создавать папку. Кому верить? Вот что
выводит колыбелька: java -cp
out/fizteh.jar:out/junit-4.11.jar:out/hamcrest-core-1.3.jar:out/asm-all-4.1.jar:out/org.json-20131017.jar
ru.fizteh.fivt.students.krivchansky.multifilemap.MultifileMapMain create
db; use db; put 1 1; get 1
created
using db
new
found
1
There should be error, if no db dir passed

+++ src/ru/fizteh/fivt/students/krivchansky/filemap/WritingUtils.java
@@ -36,7 +36,7 @@ public class WritingUtils {
 
     public void writeOffset(int offset) throws SomethingIsWrongException {
         try {
-            dataFile.writeByte(offset);
+            dataFile.writeInt(offset);
         } catch (IOException e) {
             throw new SomethingIsWrongException("Error acquired while writing to file: " + e.getMessage());
         }
